// front/src/lib/backend/apiV1/schema.d.ts
export interface components {
    schemas: {
        PostCommentModifyReqBody: {
            content: string;
        };
        RsDataVoid: {
            resultCode?: string;
            msg?: string;
            data?: unknown;
        };
        PostModifyReqBody: {
            title: string;
            content: string;
        };
        PostWriteReqBody: {
            title: string;
            content: string;
        };
        PostDto: {
            /** Format: int32 */
            id?: number;
            /** Format: date-time */
            createDate?: string;
            /** Format: date-time */
            modifyDate?: string;
            /** Format: int32 */
            authorId?: number;
            authorName?: string;
            title?: string;
            content?: string;
        };
        RsDataPostDto: {
            resultCode?: string;
            msg?: string;
            data?: components["schemas"]["PostDto"];
        };
        PostCommentWriteReqBody: {
            content: string;
        };
        PostCommentDto: {
            /** Format: int32 */
            id?: number;
            /** Format: date-time */
            createDate?: string;
            /** Format: date-time */
            modifyDate?: string;
            /** Format: int32 */
            authorId?: number;
            authorName?: string;
            /** Format: int32 */
            postId?: number;
            content?: string;
        };
        RsDataPostCommentDto: {
            resultCode?: string;
            msg?: string;
            data?: components["schemas"]["PostCommentDto"];
        };
        MemberJoinReqBody: {
            username: string;
            password: string;
            nickname: string;
        };
        MemberDto: {
            /** Format: int32 */
            id?: number;
            /** Format: date-time */
            createDate?: string;
            /** Format: date-time */
            modifyDate?: string;
            name?: string;
        };
        RsDataMemberDto: {
            resultCode?: string;
            msg?: string;
            data?: components["schemas"]["MemberDto"];
        };
        MemberLoginReqBody: {
            username: string;
            password: string;
        };
        MemberLoginResBody: {
            item?: components["schemas"]["MemberDto"];
            apiKey?: string;
            accessToken?: string;
        };
        RsDataMemberLoginResBody: {
            resultCode?: string;
            msg?: string;
            data?: components["schemas"]["MemberLoginResBody"];
        };
        MemberWithUsernameDto: {
            /** Format: int32 */
            id?: number;
            /** Format: date-time */
            createDate?: string;
            /** Format: date-time */
            modifyDate?: string;
            name?: string;
            username?: string;
        };
        AdmPostCountResBody: {
            /** Format: int64 */
            all?: number;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}